requirements.txt
appnope==0.1.2
APScheduler==3.9.1.post1
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.4.0
colorama==0.4.6
cryptography==38.0.3
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.51.0
django-apscheduler==0.6.2
django-core==1.4.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-tracking==1.5.0
entrypoints==0.4
exceptiongroup==1.0.4
executing==0.8.3
h11==0.14.0
idna==3.4
ipykernel==6.15.2
ipython==8.6.0
matplotlib-inline==0.1.6
mysqlclient==2.1.1
nest-asyncio==1.5.5
oauthlib==3.2.2
outcome==1.2.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.20
psutil==5.9.0
ptyprocess==0.7.0
Pygments==2.11.2
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
stack-data==0.2.0
tornado==6.2
traitlets==5.1.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
wcwidth==0.2.5
wincertstore==0.2
wsproto==1.2.0